requirements.txt
-e .

# aws cdk
aws-cdk.core==1.126.0
aws-cdk.aws_ec2==1.126.0
aws-cdk.aws_iam==1.126.0
aws-cdk.aws_s3_assets==1.126.0
aws-cdk.aws_elasticloadbalancingv2==1.126.0
aws_cdk.aws_elasticloadbalancingv2_targets==1.126.0
aws_cdk.aws_autoscaling==1.126.0
